+++ modules/frontend/helpers/HtmlHelper.php
@@ -2,6 +2,7 @@
 
 namespace modules\frontend\helpers;
 
+use cebe\markdown\Markdown;
 use craft\helpers\Html;
 use craft\helpers\HtmlPurifier;
 use function nl2br;
@@ -11,5 +12,13 @@ class HtmlHelper extends Html
     public static function getSafeHtml($text)
     {
         return HtmlPurifier::process(nl2br($text));
+    }
+
+    public static function getSafeMarkdownHtml($text)
+    {
+        $parser = new Markdown();
+        $html = $parser->parse($text);
+
+        return HtmlPurifier::process($html);
     }
 }