+++ modules/frontend/src/js/Shared/Layout.vue
@@ -6,25 +6,25 @@
 
                 <div class="md:flex">
                     <div class="bg-brand-900 md:flex-shrink-0 md:w-56 px-6 py-4 flex items-center">
-                        <inertia-link class="mt-1 text-white text-xl" :href="$page.siteUrl">
-                            {{ $page.siteName }}
+                        <inertia-link class="mt-1 text-white text-xl" :href="$page.props.siteUrl">
+                            {{ $page.props.siteName }}
                         </inertia-link>
                     </div>
                     <div class="px-12 py-4">
-                        <h1 class="font-bold text-3xl"> {{ $page.title }}</h1>
+                        <h1 class="font-bold text-3xl"> {{ $page.props.title }}</h1>
                     </div>
                 </div>
 
                 <div class="flex flex-grow overflow-hidden">
                     <main-menu :url="url()"
                                class="bg-brand-800 flex-shrink-0 w-56 p-12 hidden md:block overflow-y-auto"/>
-                    <div class="flex-1 overflow-hidden px-4 py-8 md:px-12 py-0 overflow-y-auto" scroll-region>
+                    <div class="flex-1 overflow-hidden px-4 py-8 md:px-12 py-0 overflow-y-auto" scroll-region="true">
 
-                        <div v-if="$page.error" class="my-2 bg-red-500 text-white px-4 py-2">
-                            {{ $page.error }}
+                        <div v-if="$page.props.error" class="my-2 bg-red-500 text-white px-4 py-2">
+                            {{ $page.props.error }}
                         </div>
-                        <div v-if="$page.notice" class="my-2 bg-green-700 text-white px-4 py-2">
-                            {{ $page.notice }}
+                        <div v-if="$page.props.notice" class="my-2 bg-green-700 text-white px-4 py-2">
+                            {{ $page.props.notice }}
                         </div>
 
                         <div class="my-8">