+++ modules/frontend/src/js/Shared/MainMenu.vue
@@ -1,8 +1,8 @@
 <template>
     <div>
         <div class="mb-4">
-            <inertia-link v-for="(item, i) in $page.mainNav" class="flex items-center group py-3" :href="item.url" :key="i">
-                <div :class="isUrl(item.url) ? 'text-white' : 'text-brand-300 group-hover:text-white'">{{ item.label }}</div>
+            <inertia-link v-for="(item, i) in $page.props.mainNav" class="flex items-center group py-3" :href="item.url" :key="i">
+                <div :class="isUrl(item.url) ? 'text-white' : 'text-brand-200 group-hover:text-white'">{{ item.label }}</div>
             </inertia-link>
         </div>
     </div>
@@ -20,7 +20,9 @@ export default {
                 return this.url === ''
             }
 
-            return urls.filter(url => this.url.startsWith(url)).length
+            let currentUrl = '/' + this.url
+
+            return urls.filter(url => currentUrl.startsWith(url)).length
         },
     },
 }