+++ modules/frontend/src/js/app.js
@@ -1,18 +1,26 @@
 import Vue from 'vue'
 import VueMeta from 'vue-meta'
-import PortalVue from 'portal-vue'
 import { InertiaApp } from '@inertiajs/inertia-vue'
 import { InertiaProgress } from '@inertiajs/progress'
+
+import '../css/app.css'
 
 Vue.config.productionTip = false
 
 Vue.use(InertiaApp)
-Vue.use(PortalVue)
 Vue.use(VueMeta)
 
-// InertiaProgress.init()
+InertiaProgress.init({
 
-let app = document.getElementById('app')
+    // The color of the progress bar.
+    color: '#1E3A8A',
+})
+
+window.baguetteBox = require('baguettebox.js');
+
+window.initLightBox = function(id) {
+    window.baguetteBox.run(id);
+}
 
 new Vue({
   metaInfo: {
@@ -24,4 +32,4 @@ new Vue({
       resolveComponent: name => import(`@/Pages/${name}`).then(module => module.default),
     },
   }),
-}).$mount(app)
+}).$mount(document.getElementById('app'))